Refactor File block to use block.json metadata
