+++ packages/block-library/src/file/block.json
@@ -1,4 +1,39 @@
 {
 	"name": "core/file",
-	"category": "common"
+	"category": "common",
+	"attributes": {
+		"id": {
+			"type": "number"
+		},
+		"href": {
+			"type": "string"
+		},
+		"fileName": {
+			"type": "string",
+			"source": "html",
+			"selector": "a:not([download])"
+		},
+		"textLinkHref": {
+			"type": "string",
+			"source": "attribute",
+			"selector": "a:not([download])",
+			"attribute": "href"
+		},
+		"textLinkTarget": {
+			"type": "string",
+			"source": "attribute",
+			"selector": "a:not([download])",
+			"attribute": "target"
+		},
+		"showDownloadButton": {
+			"type": "boolean",
+			"default": true
+		},
+		"downloadButtonText": {
+			"type": "string",
+			"source": "html",
+			"selector": "a[download]",
+			"default": "Download"
+		}
+	}
 }
